langflow-server/pyproject.toml
[project]
name = "langflow-server"
version = "1.0.0"
description = "Name My Baby - LangFlow Replacement Server"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "langchain==0.1.9",
    "langchain-openai==0.0.5",
    "openai==1.12.0",
    "pydantic==2.5.3",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

